feat: moved code around, fix copilot
