+++ internal/copilot/copilot.go
@@ -1,4 +1,4 @@
-package main
+package copilot
 
 import (
 	"context"
@@ -11,6 +11,8 @@ import (
 	"runtime"
 	"strings"
 	"time"
+
+	"github.com/charmbracelet/mods/internal/cache"
 )
 
 const (
@@ -19,8 +21,8 @@ const (
 	copilotUserAgent     = "curl/7.81.0" // Necessay to bypass the user-agent check
 )
 
-// CopilotAccessToken response from GitHub Copilot's token endpoint.
-type CopilotAccessToken struct {
+// AccessToken response from GitHub Copilot's token endpoint.
+type AccessToken struct {
 	Token     string `json:"token"`
 	ExpiresAt int64  `json:"expires_at"`
 	Endpoints struct {
@@ -37,18 +39,20 @@ type CopilotAccessToken struct {
 	} `json:"error_details,omitempty"`
 }
 
-type copilotHTTPClient struct {
+type Client struct {
 	client      *http.Client
-	AccessToken *CopilotAccessToken
+	cache       string
+	AccessToken *AccessToken
 }
 
-func newCopilotHTTPClient() *copilotHTTPClient {
-	return &copilotHTTPClient{
+func NewClient(cacheDir string) *Client {
+	return &Client{
 		client: &http.Client{},
+		cache:  cacheDir,
 	}
 }
 
-func (c *copilotHTTPClient) Do(req *http.Request) (*http.Response, error) {
+func (c *Client) Do(req *http.Request) (*http.Response, error) {
 	req.Header.Set("Accept", "application/json")
 	req.Header.Set("Editor-Version", copilotEditorVersion)
 	req.Header.Set("User-Agent", copilotUserAgent)
@@ -56,7 +60,7 @@ func (c *copilotHTTPClient) Do(req *http.Request) (*http.Response, error) {
 	isTokenExpired := c.AccessToken != nil && c.AccessToken.ExpiresAt < time.Now().Unix()
 
 	if c.AccessToken == nil || isTokenExpired {
-		accessToken, err := getCopilotAccessToken(c.client)
+		accessToken, err := c.Auth()
 		if err != nil {
 			return nil, fmt.Errorf("failed to get access token: %w", err)
 		}
@@ -124,10 +128,10 @@ func extractCopilotTokenFromFile(path string) (string, error) {
 	return "", fmt.Errorf("no token found in %s", path)
 }
 
-func getCopilotAccessToken(client *http.Client) (CopilotAccessToken, error) {
-	cache, err := NewExpiringCache[CopilotAccessToken]()
+func (c *Client) Auth() (AccessToken, error) {
+	cache, err := cache.NewExpiring[AccessToken](c.cache)
 	if err == nil {
-		var token CopilotAccessToken
+		var token AccessToken
 		err = cache.Read("copilot", func(r io.Reader) error {
 			return json.NewDecoder(r).Decode(&token)
 		})
@@ -138,12 +142,12 @@ func getCopilotAccessToken(client *http.Client) (CopilotAccessToken, error) {
 
 	refreshToken, err := getCopilotRefreshToken()
 	if err != nil {
-		return CopilotAccessToken{}, fmt.Errorf("failed to get refresh token: %w", err)
+		return AccessToken{}, fmt.Errorf("failed to get refresh token: %w", err)
 	}
 
 	tokenReq, err := http.NewRequestWithContext(context.TODO(), http.MethodGet, copilotChatAuthURL, nil)
 	if err != nil {
-		return CopilotAccessToken{}, fmt.Errorf("failed to create token request: %w", err)
+		return AccessToken{}, fmt.Errorf("failed to create token request: %w", err)
 	}
 
 	tokenReq.Header.Set("Authorization", "token "+refreshToken)
@@ -151,9 +155,9 @@ func getCopilotAccessToken(client *http.Client) (CopilotAccessToken, error) {
 	tokenReq.Header.Set("Editor-Version", copilotEditorVersion)
 	tokenReq.Header.Set("User-Agent", copilotUserAgent)
 
-	tokenResp, err := client.Do(tokenReq)
+	tokenResp, err := c.client.Do(tokenReq)
 	if err != nil {
-		return CopilotAccessToken{}, fmt.Errorf("failed to get access token: %w", err)
+		return AccessToken{}, fmt.Errorf("failed to get access token: %w", err)
 	}
 	defer func() {
 		if closeErr := tokenResp.Body.Close(); closeErr != nil && err == nil {
@@ -161,20 +165,20 @@ func getCopilotAccessToken(client *http.Client) (CopilotAccessToken, error) {
 		}
 	}()
 
-	var tokenResponse CopilotAccessToken
+	var tokenResponse AccessToken
 	if err := json.NewDecoder(tokenResp.Body).Decode(&tokenResponse); err != nil {
-		return CopilotAccessToken{}, fmt.Errorf("failed to decode token response: %w", err)
+		return AccessToken{}, fmt.Errorf("failed to decode token response: %w", err)
 	}
 
 	if tokenResponse.ErrorDetails != nil {
-		return CopilotAccessToken{}, fmt.Errorf("token error: %s", tokenResponse.ErrorDetails.Message)
+		return AccessToken{}, fmt.Errorf("token error: %s", tokenResponse.ErrorDetails.Message)
 	}
 
 	if cache != nil {
 		if err := cache.Write("copilot", tokenResponse.ExpiresAt, func(w io.Writer) error {
 			return json.NewEncoder(w).Encode(tokenResponse)
 		}); err != nil {
-			return CopilotAccessToken{}, fmt.Errorf("failed to cache token: %w", err)
+			return AccessToken{}, fmt.Errorf("failed to cache token: %w", err)
 		}
 	}
