+++ stream.go
@@ -221,7 +221,7 @@ func (m *Mods) setupStreamContext(content string, mod Model) error {
 	}
 
 	if !cfg.NoCache && cfg.cacheReadFromID != "" {
-		if err := m.cache.read(cfg.cacheReadFromID, &m.messages); err != nil {
+		if err := m.cache.Read(cfg.cacheReadFromID, &m.messages); err != nil {
 			return modsError{
 				err: err,
 				reason: fmt.Sprintf(
